refactor: helper folder renamed
