+++ tests/Challengers.UnitTests/Challengers.Application/Features/Players/Commands/DeletePlayer/DeletePlayerHandlerTests.cs
@@ -1,7 +1,7 @@
 ﻿using Challengers.Application.Features.Players.Commands.DeletePlayer;
 using Challengers.Application.Interfaces.Persistence;
 using Challengers.Domain.Entities;
-using Challengers.UnitTests.TestHelpers;
+using Challengers.UnitTests.Helpers;
 using FluentAssertions;
 using Moq;
