+++ tests/Challengers.UnitTests/Challengers.Application/Features/Players/Queries/GetPlayerById/GetPlayerByIdHandlerTests.cs
@@ -1,7 +1,7 @@
 ﻿using Challengers.Application.Features.Players.Queries.GetPlayerById;
 using Challengers.Application.Interfaces.Persistence;
 using Challengers.Domain.Entities;
-using Challengers.UnitTests.TestHelpers;
+using Challengers.UnitTests.Helpers;
 using FluentAssertions;
 using Moq;
