+++ tests/Challengers.UnitTests/Challengers.Application/Features/Tournaments/Queries/GetTournamentResult/GetTournamentResultHandlerTests.cs
@@ -2,7 +2,7 @@
 using Challengers.Application.Interfaces.Persistence;
 using Challengers.Domain.Entities;
 using Challengers.Domain.Enums;
-using Challengers.UnitTests.TestHelpers;
+using Challengers.UnitTests.Helpers;
 using FluentAssertions;
 using Moq;
